+++ R/clusters.R
@@ -69,50 +69,43 @@ update_clusterlist_df <- function(df, clusterlist) {
 
 }
 
+# FIXME
 # centroid finder for a matrix of [x, y, cluster]
-find_centroids <- function(df, total_clusters) {
-  cll <- split(df, factor(df[, 3])) #the last cluster column becomes redundant
-  l <- length(cll)
+find_centroids <- function(xyc_df, ident_levels) {
 
-  nameset <- rep(c(""), times = l)
-  xset <- rep(c(0), times = l)
-  yset <- xset
+  cll <- split(xyc_df, factor(xyc_df[, 3])) %>%
+    lapply(function(x) c(mean(x[, 1]), mean(x[, 2])))
+    
+  list_output <- init_list(ident_levels, list())
 
-  for (i in 1:l){
-    nameset[i] <- cll[[i]][,3][1]
-    xset[i] <- sum(cll[[i]][, 1]) / length(cll[[i]][, 1])
-    yset[i] <- sum(cll[[i]][, 2]) / length(cll[[i]][, 2])
+  for (ident_level in ident_levels) {
+    if (is.null(cll[[ident_level]])) next
+    list_output[[ident_level]] <- cll[[ident_level]]
   }
-  
-  list_output <- init_list(num_elements = total_clusters, init_val = list())
-  for (i in 1:l) {
-    list_output[[as.integer(nameset[i])]] <- c(xset[i], yset[i])
-  }
-  return(list_output)
+
+  unname(list_output)
 }
 
 # get reduction centroids from seurat obj, where the barcodes in the reduction
 # cell embeddings will be filtered to be exactly the same as those left in the
 # metadata incase it was additionally filtered.
-get_cluster_centroids <- function(
-  seurat_obj, reduction = "umap", passed_in_reduc_obj = FALSE
-) {
+get_cluster_centroids <- function(seurat_obj, reduction, ident_levels) {
 
-  if (passed_in_reduc_obj) {
+  if (!is_a_character(reduction)) {
     reduc_coords <- reduction@cell.embeddings[, 1:2]
   } else {
     reduc_coords <- get_2d_embedding(seurat_obj, reduction)
   }
 
   find_centroids(
-    df = data.frame(
+    xyc_df = data.frame(
       rcppFilterReductionCoords(
         seuratBarcodes = rownames(seurat_obj@meta.data),
         reductionCoords = reduc_coords
       ),
-      seurat_obj@meta.data[["seurat_clusters"]]
+      "ident" = seurat_obj@meta.data[["seurat_clusters"]]
     ),
-    total_clusters = get_num_total_clusters(seurat_obj)
+    ident_levels
   )
 }