+++ tests/testthat/test-ApotcData.R
@@ -1,11 +1,21 @@
+# test_that("ident_into_seurat_clusters works", {
+
+# })
+
+data("combined_pbmc")
 non_subset_apotc_data <- ApotcData(
-	seurat_obj = get(data("combined_pbmc")),
+	seurat_obj = combined_pbmc,
+	alt_ident = NULL,
 	metadata_filter_condition = "",
 	clonecall = "CTstrict",
 	reduction_base = "umap",
 	clone_scale_factor = 0.300051,
 	rad_scale_factor = 0.95
 )
+
+test_that("have_default_idents work", {
+	expect_true(have_default_idents(non_subset_apotc_data))
+})
 
 expect_common_apotc_els_equal <- function(test_apotc_data) {
 
@@ -15,7 +25,7 @@ expect_common_apotc_els_equal <- function(test_apotc_data) {
 	)
 	expect_equal(test_apotc_data@clonecall, non_subset_apotc_data@clonecall)
 	
-	expect_equal(test_apotc_data@num_clusters, non_subset_apotc_data@num_clusters)
+	expect_equal(get_num_clusters(test_apotc_data), get_num_clusters(non_subset_apotc_data))
 
 	expect_equal(
 		test_apotc_data@clone_scale_factor, non_subset_apotc_data@clone_scale_factor
@@ -51,7 +61,7 @@ test_that("The default case ApotcData constructor works", {
 		getdata("get_clone_sizes", "raw_strict_clone_sizes")
 	)
 
-	expect_equal(test_apotc_data@num_clusters, 17)
+	expect_equal(get_num_clusters(test_apotc_data), 17)
 
 	expect_equal(test_apotc_data@clone_scale_factor, 0.300051)
 	expect_equal(test_apotc_data@rad_scale_factor, 0.95)
@@ -67,10 +77,28 @@ test_that("The default case ApotcData constructor works", {
 
 })
 
+test_that("changing active ident affects clustering", {
+
+	test_apotc <- ApotcData(
+		seurat_obj = combined_pbmc,
+		alt_ident = "seurat_clusters",
+		metadata_filter_condition = "",
+		clonecall = "CTstrict",
+		reduction_base = "umap",
+		clone_scale_factor = 0.300051,
+		rad_scale_factor = 0.95
+	)
+
+	expect_common_apotc_els_equal(test_apotc)
+
+	# TODO more tests on custom ident
+})
+
 test_that("The subset case ApotcData constructor works", {
 
 	test_apotc_data <- ApotcData(
-		seurat_obj = get(data("combined_pbmc")),
+		seurat_obj = combined_pbmc,
+		alt_ident = NULL,
 		metadata_filter_condition = "seurat_clusters != 1",
 		clonecall = "CTstrict",
 		reduction_base = "umap",
@@ -101,7 +129,8 @@ test_that("The subset case ApotcData constructor works", {
 test_that("ApotcData subsetting works for 1 seurat cluster", {
 
 	test_apotc_data <- ApotcData(
-		seurat_obj = get(data("combined_pbmc")),
+		seurat_obj = combined_pbmc,
+		alt_ident = NULL,
 		metadata_filter_condition = "seurat_clusters == 4",
 		clonecall = "CTstrict",
 		reduction_base = "umap",
@@ -168,7 +197,8 @@ test_that("circlepackClones packs right for the default case", {
 test_that("circlepackClones works for the subset case", {
 
 	test_apotc_data <- ApotcData(
-		seurat_obj = get(data("combined_pbmc")),
+		seurat_obj = combined_pbmc,
+		alt_ident = NULL,
 		metadata_filter_condition = "seurat_clusters != 1",
 		clonecall = "CTstrict",
 		reduction_base = "umap",
@@ -209,7 +239,8 @@ test_that("circlepackClones packs right for the subset case", {
 	skip_on_cran()
 	
 	test_apotc_data <- ApotcData(
-		seurat_obj = get(data("combined_pbmc")),
+		seurat_obj = combined_pbmc,
+		alt_ident = NULL,
 		metadata_filter_condition = "seurat_clusters != 1",
 		clonecall = "CTstrict",
 		reduction_base = "umap",