+++ tests/testthat/testdata/v0/SCIP.R
@@ -1,6 +1,0 @@
-data("mini_clonotype_data", "mini_seurat_obj")
-
-test_pbmc <- mini_seurat_obj
-test_tcr <- mini_clonotype_data
-
-rm("mini_clonotype_data", "mini_seurat_obj")